refactor: move tests to tests folder
